requirements.txt
fastapi
uvicorn
yt-dlp
requests
python-dotenv